Update & move create & import wallet pages
