--- src/pages/Wallets/routes/Create/index.js
@@ -1,8 +1,0 @@
-// @flow
-
-import View from './WalletsCreateViewContainer'
-
-export default {
-  path: 'create',
-  component: View,
-}
