--- src/pages/Wallets/routes/Import/index.js
@@ -1,8 +1,0 @@
-// @flow
-
-import View from './WalletsImportViewContainer'
-
-export default {
-  path: 'import',
-  component: View,
-}
